+++ samplesheets/tests/test_models.py
@@ -109,6 +109,7 @@ DEFAULT_DESCRIPTION = 'Description'
 DEFAULT_COMMENTS = {'comment': 'value'}
 
 ISATAB_DATA = {'i_investigation.txt': '', 's_study.txt': '', 'a_assay.txt': ''}
+ISATAB_DESC = 'description'
 
 PLUGIN_NAME_DNA_SEQ = 'samplesheets_assay_dna_sequencing'
 PLUGIN_NAME_GENERIC_RAW = 'samplesheets_assay_generic_raw'
@@ -357,6 +358,7 @@ class SampleSheetModelMixin:
         parser_version=None,
         user=None,
         extra_data={},
+        description=None,
     ):
         """Create an ISATab object in the database"""
         values = {
@@ -368,6 +370,7 @@ class SampleSheetModelMixin:
             'parser_version': parser_version,
             'user': user,
             'extra_data': extra_data,
+            'description': description,
         }
         obj = ISATab(**values)
         obj.save()
@@ -1263,6 +1266,7 @@ class TestISATab(TestSampleSheetBase):
             tags=[],
             parser_version=DEFAULT_PARSER_VERSION,
             user=self.user_owner,
+            description=ISATAB_DESC,
         )
 
     def test_initialization(self):
@@ -1277,6 +1281,7 @@ class TestISATab(TestSampleSheetBase):
             'parser_version': DEFAULT_PARSER_VERSION,
             'user': self.user_owner.pk,
             'extra_data': {},
+            'description': ISATAB_DESC,
             'sodar_uuid': self.isatab.sodar_uuid,
         }
         self.assertEqual(model_to_dict(self.isatab), expected)
@@ -1312,7 +1317,7 @@ class TestISATab(TestSampleSheetBase):
                 '%Y-%m-%d %H:%M:%S'
             ),
         )
-        self.assertEqual(self.isatab.get_name(), expected)
+        self.assertEqual(self.isatab.get_full_name(), expected)
 
     def test_get_name_no_title(self):
         """Test get_name() with no title"""
@@ -1324,7 +1329,7 @@ class TestISATab(TestSampleSheetBase):
                 '%Y-%m-%d %H:%M:%S'
             ),
         )
-        self.assertEqual(self.isatab.get_name(), expected)
+        self.assertEqual(self.isatab.get_full_name(), expected)
 
     def test_get_name_no_archive(self):
         """Test get_name() with no title or archive name"""
@@ -1337,7 +1342,7 @@ class TestISATab(TestSampleSheetBase):
                 '%Y-%m-%d %H:%M:%S'
             ),
         )
-        self.assertEqual(self.isatab.get_name(), expected)
+        self.assertEqual(self.isatab.get_full_name(), expected)
 
 
 class TestIrodsAccessTicket(TestSampleSheetBase):