+++ samplesheets/views.py
@@ -13,11 +13,13 @@ import zipfile
 
 from django.conf import settings
 from django.contrib import messages
+from django.db.models.functions import Now
 from django.http import HttpResponse
 from django.shortcuts import redirect
 from django.urls import reverse
 from django.utils.safestring import mark_safe
 from django.utils.text import slugify
+from django.utils.timezone import localtime
 from django.views.generic import (
     DeleteView,
     FormView,
@@ -55,6 +57,7 @@ from samplesheets.forms import (
     IrodsRequestForm,
     IrodsRequestAcceptForm,
     SampleSheetVersionCompareForm,
+    SampleSheetVersionEditForm,
 )
 from samplesheets.io import (
     SampleSheetIO,
@@ -334,7 +337,7 @@ class SampleSheetImportMixin:
         if ui_mode:
             success_msg = '{}d sample sheets from {}'.format(
                 action.capitalize(),
-                'version {}'.format(isa_version.get_name())
+                'version {}'.format(isa_version.get_full_name())
                 if action == 'restore'
                 else 'ISA-Tab import',
             )
@@ -523,7 +526,7 @@ class SampleSheetISAExportMixin:
         else:
             file_name = clean_sheet_dir_name(project.title)
         if isa_version:
-            file_name += '_' + isa_version.date_created.strftime(
+            file_name += '_' + localtime(isa_version.date_created).strftime(
                 '%Y-%m-%d_%H%M%S'
             )
             if isa_version.user:
@@ -551,7 +554,9 @@ class SampleSheetISAExportMixin:
             )
             if isa_version:
                 tl_event.add_object(
-                    obj=isa_version, label='isatab', name=isa_version.get_name()
+                    obj=isa_version,
+                    label='isatab',
+                    name=isa_version.get_full_name(),
                 )
 
         # Initiate export
@@ -1545,14 +1550,13 @@ class SampleSheetVersionListView(
     def get_queryset(self):
         return ISATab.objects.filter(
             project__sodar_uuid=self.kwargs['project']
-        ).order_by('-pk')
+        ).order_by('-date_created')
 
     def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(*args, **kwargs)
         context['current_version'] = None
         choices = [
-            (str(i.sodar_uuid), i.date_created.strftime('%Y-%m-%d %H:%M:%S'))
-            for i in self.get_queryset()
+            (str(obj.sodar_uuid), obj.get_name()) for obj in self.get_queryset()
         ]
         context['sheet_select_form'] = SampleSheetVersionCompareForm(
             choices=choices
@@ -1678,7 +1682,9 @@ class SampleSheetVersionRestoreView(
                 description='restore sheets from version {isatab}',
             )
             tl_event.add_object(
-                obj=isa_version, label='isatab', name=isa_version.get_name()
+                obj=isa_version,
+                label='isatab',
+                name=isa_version.get_full_name(),
             )
 
         try:
@@ -1708,12 +1714,45 @@ class SampleSheetVersionRestoreView(
             # Edit isa_version to bump it in the list
             if 'RESTORE' not in isa_version.tags:
                 isa_version.tags.append('RESTORE')
+            isa_version.date_created = Now()
             isa_version.save()
 
         return redirect(
             reverse(
                 'samplesheets:project_sheets',
                 kwargs={'project': project.sodar_uuid},
+            )
+        )
+
+
+class SampleSheetVersionUpdateView(
+    LoginRequiredMixin,
+    LoggedInPermissionMixin,
+    ProjectPermissionMixin,
+    InvestigationContextMixin,
+    UpdateView,
+):
+    """Sample sheet version update view"""
+
+    permission_required = 'samplesheets.manage_sheet'
+    model = ISATab
+    form_class = SampleSheetVersionEditForm
+    template_name = 'samplesheets/version_update.html'
+    slug_url_kwarg = 'isatab'
+    slug_field = 'sodar_uuid'
+
+    def form_valid(self, form):
+        obj = form.save()
+        messages.success(
+            self.request,
+            'Description updated for sheet version "{}".'.format(
+                obj.get_full_name()
+            ),
+        )
+        return redirect(
+            reverse(
+                'samplesheets:versions',
+                kwargs={'project': self.get_project().sodar_uuid},
             )
         )
 
@@ -1756,11 +1795,15 @@ class SampleSheetVersionDeleteView(
                 status_type='OK',
             )
             tl_event.add_object(
-                obj=self.object, label='isatab', name=self.object.get_name()
+                obj=self.object,
+                label='isatab',
+                name=self.object.get_full_name(),
             )
         messages.success(
             self.request,
-            'Deleted sample sheet version: {}'.format(self.object.get_name()),
+            'Deleted sample sheet version: {}'.format(
+                self.object.get_full_name()
+            ),
         )
         return reverse(
             'samplesheets:versions', kwargs={'project': project.sodar_uuid}
@@ -1890,7 +1933,7 @@ class IrodsAccessTicketDeleteView(
     InvestigationContextMixin,
     DeleteView,
 ):
-    """Sample sheet version deletion view"""
+    """iRODS access ticket deletion view"""
 
     permission_required = 'samplesheets.delete_sheet'
     template_name = 'samplesheets/irodsaccessticket_confirm_delete.html'